#364 counter-increment support for footnotes.
Still struggling with correct counter-reset behavior.
