+++ openhtmltopdf-core/src/main/java/com/openhtmltopdf/layout/counter/AbstractCounterContext.java
@@ -1,0 +1,11 @@
+package com.openhtmltopdf.layout.counter;
+
+import java.util.List;
+
+public interface AbstractCounterContext {
+
+    int getCurrentCounterValue(String name);
+
+    List<Integer> getCurrentCounterValues(String name);
+
+}
